ordem-service/requirements.txt
uvicorn[standard]==0.15.0
fastapi==0.68.1
pydantic!=1.8
aio_pika==8.1.1
redis_om==0.0.27
asyncio==3.4.3
opentelemetry-sdk==1.12.0rc2
opentelemetry-exporter-jaeger-thrift==1.12.0rc2
opentelemetry-instrumentation-fastapi==0.32b0
opentelemetry-sdk==1.12.0rc2
opentelemetry-instrumentation-redis==0.32b0
opentelemetry-exporter-otlp-proto-grpc==1.12.0rc2
opentelemetry-instrumentation-aio-pika==0.32b0
aiohttp==3.8.1
opentelemetry.instrumentation.aiohttp_client==0.32b0
starlette_prometheus